# docs/openapi.yaml
openapi: 3.0.3
info:
  title: BanHang Backend API
  version: 2.0.0
  description: Versioned NestJS backend API for auth, CRM, content, catalog, cart, orders, payments, inventory, shipping, notifications, reporting, and store locator flows.
servers:
  - url: http://localhost:4000/api/v1
...
paths:
...
  /account:
    delete:
      tags: [Account]
      summary: Delete and anonymize the current customer account
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeleteAccountRequest'
      responses:
        '200':
          description: Account anonymized and personal side data removed
  /account/dashboard:
    get:
      tags: [Account]
      summary: Customer dashboard summary
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Dashboard summary
  /account/orders:
    get:
      tags: [Account]
      summary: List current customer orders with filter and search
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: q
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
            enum: [created, confirmed, shipping, completed, canceled, returned]
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 50
      responses:
        '200':
          description: Paged customer order list
  /account/orders/{id}/reorder:
    post:
      tags: [Account]
      summary: Re-add a previous order into the current cart
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Reorder result with added and skipped items
  /account/loyalty:
    get:
      tags: [Account]
      summary: Loyalty summary and earning history derived from completed orders and referral rewards
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Loyalty points balance, tier, and history
  /account/loyalty/redeem:
    post:
      tags: [Account]
      summary: Redeem loyalty points into a single-use coupon voucher
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RedeemLoyaltyRequest'
      responses:
        '200':
          description: Loyalty redemption result with created coupon
  /account/referral:
    get:
      tags: [Account]
      summary: Get current customer referral code, share link, and referral stats
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Referral summary with signups and rewards
  /account/referral/regenerate:
    post:
      tags: [Account]
      summary: Rotate current customer referral code and return updated referral summary
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Referral summary with refreshed code
  /account/profile:
    get:
      tags: [Account]
      summary: Get current customer profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Customer profile
    patch:
      tags: [Account]
      summary: Update current customer profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Customer profile updated
  /account/addresses:
    get:
      tags: [Account]
      summary: List address book
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Address book
    post:
      tags: [Account]
      summary: Create address
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAccountAddressRequest'
      responses:
        '200':
          description: Address book updated
  /account/addresses/{id}:
    patch:
      tags: [Account]
      summary: Update address
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAccountAddressRequest'
      responses:
        '200':
          description: Address book updated
    delete:
      tags: [Account]
      summary: Delete address
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Address book updated
  /account/addresses/{id}/default:
    post:
      tags: [Account]
      summary: Set default address
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Address book updated
  /account/export:
    get:
      tags: [Account]
      summary: Export current customer data
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Customer data export
...
  /orders:
    get:
      tags: [Orders]
      summary: List orders for current user or all orders for privileged users
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Order list
    post:
      tags: [Orders]
      summary: Alias of checkout endpoint
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckoutRequest'
      responses:
        '201':
          description: Order created
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
  schemas:
...
    UpdateProfileRequest:
      type: object
      properties:
        fullName:
          type: string
          minLength: 2
        phone:
          type: string
          minLength: 8
    RedeemLoyaltyRequest:
      type: object
      required: [points]
      properties:
        points:
          type: integer
          minimum: 500
    DeleteAccountRequest:
      type: object
      required: [password]
      properties:
        password:
          type: string
          minLength: 8
        reason:
          type: string
          minLength: 3
    CreateAccountAddressRequest:
      type: object
      required: [fullName, phone, district, addressLine]
      properties:
        label:
          type: string
        fullName:
          type: string
          minLength: 2
        phone:
          type: string
          minLength: 8
        province:
          type: string
        district:
          type: string
          minLength: 2
        ward:
          type: string
        addressLine:
          type: string
          minLength: 4
        country:
          type: string
        isDefault:
          type: boolean
    UpdateAccountAddressRequest:
      type: object
      properties:
        label:
          type: string
        fullName:
          type: string
          minLength: 2
        phone:
          type: string
          minLength: 8
        province:
          type: string
        district:
          type: string
          minLength: 2
        ward:
          type: string
        addressLine:
          type: string
          minLength: 4
        country:
          type: string
        isDefault:
          type: boolean
...
    CheckoutRequest:
      type: object
      required: [reservationId, paymentMethod, shippingMethod]
      properties:
        reservationId:
          type: string
        addressId:
          type: string
        address:
          type: object
          required: [receiverName, phone, district, addressLine, country]
          properties:
            receiverName:
              type: string
            phone:
              type: string
            province:
              type: string
            district:
              type: string
            ward:
              type: string
            addressLine:
              type: string
            country:
              type: string
        saveAddress:
          type: boolean
        paymentMethod:
          type: string
          enum: [cod, vnpay, momo, zalopay, stripe, paypal, bank_transfer]
        shippingMethod:
          type: string
          enum: [standard, express, same_day, pickup]
        notes:
          type: string
...